requirements.txt
tkinter
robloxpy
pillow
pyinstaller